requirements.txt
torch
torchvision
torchaudio
numpy
matplotlib
scikit-learn
seaborn
tqdm
Pillow